refactor out files to keep it clean
